requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.3.0
SQLAlchemy==2.0.0
python-dotenv==1.0.0
pymysql==1.1.0